+++ dacite/CHANGELOG.md
@@ -6,6 +6,7 @@
  - Extension `VK_AMD_shader_trinary_minmax` (revision 1)
  - Extension `VK_AMD_shader_explicit_vertex_parameter` (revision 1)
  - Extension `VK_AMD_gcn_shader` (revision 1)
+ - Extension `VK_NV_dedicated_allocation` (revision 1)
 
 ### Breaking changes